[TerriaJS/aremi-natmap#136] PointParameterEditor/PointParameter: Refactor to push validation and parsing into PointParameter.
