+++ lib/Models/PointParameter.js
@@ -1,11 +1,14 @@
 'use strict';
 
 /*global require*/
+var Cartographic = require('terriajs-cesium/Source/Core/Cartographic');
 var CesiumMath = require('terriajs-cesium/Source/Core/Math');
+var defined = require('terriajs-cesium/Source/Core/defined');
 var defined = require('terriajs-cesium/Source/Core/defined');
 var defineProperties = require('terriajs-cesium/Source/Core/defineProperties');
 var FunctionParameter = require('./FunctionParameter');
 var inherit = require('../Core/inherit');
+var knockout = require('terriajs-cesium/Source/ThirdParty/knockout');
 var defaultValue = require('terriajs-cesium/Source/Core/defaultValue');
 
 /**
@@ -24,6 +27,25 @@ var defaultValue = require('terriajs-cesium/Source/Core/defaultValue');
  */
 var PointParameter = function(options) {
     FunctionParameter.call(this, options);
+
+    /**
+     * Determines whether the point parameter is currently holding valid data.
+     * @member {*} isValid
+     * @memberof FunctionParameter.prototype
+     */
+    knockout.defineProperty(this, 'isValid', {
+        get: function () {
+            if (defined(this.value)) {
+                if (this.value instanceof Cartographic) {
+                    if ((isNaN(this.value.longitude) === false) && (isNaN(this.value.latitude) === false)) {
+                        return true;
+                    }
+                }
+            }
+
+            return false;
+        }
+    });
 };
 
 inherit(FunctionParameter, PointParameter);
@@ -89,6 +111,32 @@ PointParameter.prototype.getGeoJsonFeature = function(value) {
             coordinates: coordinates
         }
     };
+};
+
+/**
+ * Parses the value into a Cartographic and stores the value as this.value, otherwise updates this.value to undefined.
+ * @param {String} value Text that is to be parsed.
+ */
+PointParameter.prototype.parseParameterLonLat = function(value) {
+    this.value = this.tryParseCartographicValueFromTextLonLat(value);
+}
+
+/**
+ * Parses the value into a Cartographic if it is in a valid format, otherwise returns undefined.
+ * @param {String} value Text that is to be parsed. The value is expected to be formatted as "{Lon},{Lat}".
+ * @return {Cartographic} The value if it was able to be parsed otherwise undefined.
+ */
+PointParameter.prototype.tryParseCartographicValueFromTextLonLat = function(value) {
+    const coordinates = value.split(',');
+    if (coordinates.length >= 2) {
+        const lon = parseFloat(coordinates[0]);
+        const lat = parseFloat(coordinates[1]);
+        if (!isNaN(lat) && !isNaN(lon)) {
+            return Cartographic.fromDegrees(lon, lat);
+        }
+    }
+
+    return undefined;
 };
 
 PointParameter.formatValueAsString = PointParameter.prototype.formatValueAsString;
